pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harness-workflow"
version = "0.1.0"
description = "Harness Engineering workflow CLI and installable Codex skill."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "questionary>=2.0.0",
  "pyyaml>=6.0",
]
authors = [
  { name = "togally" }
]
license = { text = "MIT" }
keywords = ["harness", "agent", "workflow", "codex", "claude"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.scripts]
harness = "harness_workflow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
harness_workflow = [
  "skills/harness/SKILL.md",
  "skills/harness/agent/*.md",
  "assets/skill/SKILL.md",
  "assets/skill/agents/openai.yaml",
  "assets/skill/assets/templates/*.tmpl",
  "assets/templates/*.md",
  "assets/skill/references/*.md",
  "assets/skill/scripts/*.py",
  "assets/skill/tests/*.py",
  "assets/scaffold_v2/WORKFLOW.md",
  "assets/scaffold_v2/.workflow/context/*.md",
  "assets/scaffold_v2/.workflow/context/roles/*.md",
  "assets/scaffold_v2/.workflow/context/roles/directors/*.md",
  "assets/scaffold_v2/.workflow/context/project/*.md",
  "assets/scaffold_v2/.workflow/context/team/*.md",
  "assets/scaffold_v2/.workflow/context/experience/index.md",
  "assets/scaffold_v2/.workflow/context/experience/roles/*.md",
  "assets/scaffold_v2/.workflow/context/experience/tool/*.md",
  "assets/scaffold_v2/.workflow/context/experience/risk/*.md",
  "assets/scaffold_v2/.workflow/state/experience/*.md",
  "assets/scaffold_v2/.workflow/constraints/*.md",
  "assets/scaffold_v2/.workflow/evaluation/*.md",
  "assets/scaffold_v2/.workflow/flow/*.md",
  "assets/scaffold_v2/.workflow/tools/*.md",
  "assets/scaffold_v2/.workflow/tools/catalog/*.md",
  "harness_workflow/tools/*.py",
]
